Cleanup and added kcontext_copy().
Adding query evaluation for beliefs is not that easy:

Firstly, we could either simply apply all operations to all setups at
different plausibility levels (e.g., add a split literal to all of
them), or we could try to always only consider the most plausible setup
and apply respective operations to the next-plausible if the other one
gets inconsistent.

Secondly, when actions are added, H+ may change. Then, and only then,
the static setup must be recomputed. Actually, that's not a problem,
because this change can only be due to more variables or new names in
the query, i.e., the change cannot be due to the static setup. Thus it
should be fine to use the new static setup to compute the belief ranking
and add each plausibilities setup to a copy of the new static setup.