Added belief contexts.
More precisely, context_t now is either a knowledge or a belief context.
Also moved query_test_clause() to setup.c and renamed it
bsetup_with_splits_and_sf_subsumes(). This function collects the PEL of
the given clause and its to-be-split SF literals. As a side-effect, it
doesn't eat up its PEL parameter anymore.