Using non-const iterators where appropriate.
Setups change (not semantically) when they are minimized.
For that reason, some [b]setups need to be non-const although.
A bsetup is a vector of setups.
We used const iterators and then val_unsafe to access the elements and
hand them over to the setup functions. Now, we use normal iterators and
use non-const val. That, by the way, was declared const accidentally
before.