Changed kitchen BAT for demo.
Performance is very bad.
The resulting setups are huge and grow linearly in the number of
actions, and query evaluation seems to grow quadratically in the size
of the setup.

The split optimizations in setup are still worth their cost (replacing
setup_clause_small_pel() with setup_full_pel() costs about 20% and
removing the expensive part of setup_relevant_split() costs 1000%).