Fixed: Setup::GroundBoxes(z) didn't ground for z.
I'm not sure if that really was too bad, because for SSAs, it suffices
to ground for all prefixes of z, because the SSAs do the last step to
situation z.

But not grounding for z leads to weird behavior with SFs: if you add a
sensing result [z]SF(a), this only takes effect once you have grounded
for z. That is, after AddSensingResult(z, a, true), Entail({[z]SF(a)})
fails. This doesn't occur in practice, because we usually perform
queries about Entails({[z.a]A}), which leads to calls GroundBoxes(z.a)
and hence z was grounded.

For performance reasons one may thus want to stick with just grounding
the prefixes of z.