Fixed: PEL(c) contains dynamic literals from the query clause.
Suppose p is a fluent with an SSA and [z]p is a candidate for PEL(c).
Splitting [z]p is redundant because we can split p as well and obtain
[z](~)p by unit propagation with the SSAs.

But if [z]p has no SSA, we may need to split it, because it may be part
of a tautology, like ([z]p v [z]~p). Then [z]p shall be in PEL(c)
because [z]p and [z]~p are in REL(c).