There is a unary and binary Maybe now.
Maybe is a variadic template (without any content) with two
specializations for the unary and binary case.

In the SimpleClause and Clause unification we return the MGU and the
result of unification together. The first solution was to use
Maybe<std::pair<Unifier, [Simple]Clause>>. Now it's
Maybe<Unifier, [Simple]Clause>>.