Refactored Formula, introduced BasicActionTheory class.
The DynamicAxioms class is now called BasicActionTheory and also
interfaces the Setup or Setups. Thus we don't need double methods in
Formula. In fact, the Formula::AddToSetup[s]() and Formula::Eval() are
now BasicActionTheory::Add() and BasicActionTheory::Entails().

Not sure if that's really nice, but probably better than the old
solution.