Fixed: conditional propagation in belief setups.
Setups::PropagateBeliefs() goes through the conditionals and adds
neg_phi_or_psi to each setup until it finds one that does not entail
neg_phi, that is, is consistent with phi.

So far so good.

The procedure added a new, clean setup to the end until m+1 there
are setups where m is the number of conditionals.
But we don't need all of them in all cases, for example, we may have
ss_[0] != ss_[1] != ... != ss_[m-1] == ss_[m].
In this case, we need to take care we haven't added an empty setup
ss_[m]: we want to have ss_[m] == ss_[m-1], or just leave out ss_[m]
for efficiency. The fix implements the latter.