Grounder avoids regrounding after AddClause(), PrepareFor().
When no new names were added in AddClause() and PrepareFor(), which
would make (at least partially) regrounding of old clauses necessary,
then Ground() creates a new setup that inherits from the previous
result of Ground() and adds only the new clause.

The minesweeper example now works. So far it's just translation of the
old one to the new library.