Replaced incr_iterator with int_iterator.
int_iterator takes an (integer) type and a function which is applied to
it. The only difference to the incr_iterator is that the function is
not nullary but applied to the index.

Also changed the order of the UnaryFunction / UnaryPredicate arguments
in iterators: it's now always the last argument. That works better with
default arguments UnaryFunction() / UnaryPredicate().