Difference between old and new minesweeper is bug in old one.
For ./minesweeper 8 8 10 1, after the initial opening the new
implementation goes on to explore (0,2). The old implementation misses
that one because for some reason it misses that the setup subsumes (in
fact, contains) the clause [~Mine(0,2)]. Instead, the old
implementation finds that (1,2) contains a mine. It is an equally good
move in this particular game, but in general the missing subsumption is
a bug of course. I haven't bothered looking why that is, but the bug
clearly is in the old implementation (and its more complex Setup class).