First version of Grounder::AssignTerms().
An open question is how (and if at all) to determine
Grounder::RelevantAssignTerms().

There's also a bug in the splitting and assigning of terms regarding
the query: when setting t=n, we must not substitute n for t in the
query, because t may not be ground yet. Instead we could wait when we
check for clause subsumption.