Cleaned up copy/move constructors/assignment.
When copy (move) constructor/assignment is deleted, then move (copy)
constructor/assignment is implicitly also deleted. But it should be
fine to move a Grounder or a Solver or a Factory; so their move
constructor/assignment are now declared as default.

Also, some of the declarations were just wrong or redundant.