Fixed: Grounder::AssignLiterals() missed literals.
It missed that when (t != n) is occurs somewhere, we need to consider
(t = x) as split literal, not just (t = n). This fixes the bug
mentioned in commit a2b02e99.

Furthermore, more batch-assign-literals must be considered. When adding
P(#1), P(#2), P(#3) to the setup, we also need to consider the
assignment P(x), not just P(n) individually for single n.