Rewrote/added Grounder::[Relevant]LiteralAssignments().
The semantics for multi-assignment is not that l^x_n is added for all
\vec{n}, but rather a clause with ewff. For instance, this happens in
test-introspection.lela, where we have the set
  {p(m) = T, p(m) = #10, p(n) = T}
which misses p(m) = T and p(n) = #10, i.e., it does not correspond to
p(x) = y, but rather to (x != n || y != #10) -> p(x) = y.

The reason is that we build the multi-assignment sets using
Term::Isomorphic().