New web-interface for Minesweeper.
The performance was severely limited by the hacky visualization. The new
one uses div's, which are updated individually. There seems to be a bug
in Emscripten that makes mw-js.js crash whenever some more string
handling is done in PlayTurn(). This is why the interface misses some
information (in particular, timing results) at the moment.