Fixed: missing lhs in Setup::Units::Determines().
Setup::Units::Determines() didn't check if the lhs of the literal found
in the hashtable's bucket actually agrees with the one we are looking
for. This bug could lead to wrong behaviour of Supet::Determines() and
hence also in Solver::Split() and thus to wrong reasoning results.

Setup::Determines() now returns the name (or null term).