Minesweeper optionally uses Solver::Determines() now.
This requires a slightly different model, as we need to represent that
a cell contains a mine with a literal [t = F] instead of [t /= T], so
that Solver::Determines() can find non-mine cells. This version is
activated by defining USE_DETERMINES. It appears to be faster, so it's
activated at the moment (in CMakeLists).