Complete rewrite of Grounder.
The new grounder closely follows the procedure in the extended arXiv
version of the IJCAI-17 paper, except that names for quantifiers in
the query are pre-allocated.

Consistency guarantees are not yet considered, which also causes some
guarantee-related unit tests to fail. This is awkward since the
grounder knowing about consistency should only affect performance, not
semantics.

Also dropped the convention that non-const getters return a pointer:
it's too annoying. For out- or in/out-parameters the pointer convention
remains.