KnowledgeBase handles G now.
G is now handled as part of the Representation Theorem procedure in
KnowledgeBase, as proposed in the IJCAI paper.

Before this commit, a Boolean assume_consistent was managed and handed
over to the Solver, which then called Grounder::GuaranteeConsistancy(),
which led to many redundant calls.

The flag in Solver still exists for now, as unit tests use it.
Otherwise it is of no use and could be removed (which would slightly
clarify code in Solver).