Fixed: KnowledgeBase::ReduceModalities() and G in objective formulas.
Not sure if there is any use-case for this, but it shouldn't crash.
The unit test is: G false.

The paper says G does not affect the objective setup. However,
behaviour is undefined if the guarantee doesn't hold, and if it is, we
can safely reduce the objective setup as well. So it's safe to differ
from the theory in this point.