Fixed: Setup::Units::Resize(n) is well-defined for n = 0.
This is relevant for the last Grounder::Ply. We create the shallow copy
for the setup first and then minimize the setup, which could increase
Setup::Units::n_orig_ > 0, whereas ShallowCopy::n_units_ = 0.

In general, one should call ShallowCopy::Minimize() instead of
Setup::Minimize(), and ShallowCopy::Minimize() doesn't
Setup::Units::SealOriginalUnits(). In the special case of a ShallowCopy
of an empty setup, however, Setup::Minimize() should be a safe call and
now is.