Added ActionTheory::[Rewrite|Regress]().
ActionTheory::Rewrite() turns post-action symbols into new symbols and
manages this mapping. What's still needed is to compile the dynamic
axioms into proper+ form.

ActionTheory::Regression() works as usual with the modification that
a literal may unify with >= 0 left-hand sides of axioms; it is
substituted with every corresponding right-hand side. That way, we
often get simpler axioms and can avoid additional quantifiers, e.g.,
when we want to simulat the propositional case.