Re-implemented consistency guarantee.
The new consistency guarantee does not build a new, minimised setup,
but instead manages indices of relevant terms and clauses. Essentially
the idea is that we have a graph where terms are connected through
clauses.

It's faster than before, but it's still slow in cases like Minesweeper.
The problem here is that clauses that were relevant are kept relevant
and hence are split.

The new implementation deviates from the semantics in the IJCAI paper,
as subsumed clauses might still connect literals. Could and should we
work around this, or rather change the semantics?