Setup::[Minimize|clause]() use unit-subsumption.
Setup::Minimize() minimised with respect to subsumption, which costs
n^2. Now we only check for subsumption by unit clauses.

Setup::clause() now skips clauses subsumed by unit clauses.

To make all this possible with minimal overhead,
Clause::PropagateUnit[s]() integrates unit subsumption.

This means a departure from the IJCAI semantics of UP^-, in that it now
minimises with respect to unit subsumption only and not full
subsumption. The rationale behind unit subsumption is that it is more
`limited' as minimising a whole set of clauses is a relatively expensive
operation. The difference is probably not very important in practical
applications, because one would rarely add subsuming non-unit clauses to
a knowledge base (whereas splits and sensing literals are examples for
subsuming unit clauses).