Rewrote Grounder::CloseRelevance().
The new implementation searches from new clauses and clauses that
contain new relevant query terms for relevant clauses, which renders the
clauses connected to the origin clause as relevant.

Similarly, to determine clauses as irrelevant, it looks from those
clauses that contain term from a new unit clauses for clauses that
contain a relevant query term; if no such clause can be found, the
clauses connected to the origin are irrelevant.