Revised memory layout of Terms.
Terms now not only store whether they are name but also whether they are
primitive, a variable, or anything else (i.e., non-ground function or
even a non-well-formed Term) directly in their representation. This
allows for faster access to these properties. The full representations
are now stored in distinct heaps for every class. In particular for
primitive Terms this means that indices are smaller. Hopefully we can
use this later to represent a mapping from [primitive|name] Terms to
other objects as an array.