Added name order. Cosmetic changes.
Added ActivityOrder for names, one per function. So far it's not faster
than CandidateName(), but it's likely that the current bumping isn't
very smart.

Fixed an assertion: a clause may be watched more than one time in by
the same function. That's not intended, but possible because the
current implementation lazily removes watchers. This might be worth
looking into.

Simplified DenseMap API and some other minor things.