Simplified Term, Literal.
Terms and Literals are now assumed to be ground and well-formed.
The idea is that non-ground and non-well-formed terms and literals only
exist on formula-level, whereas the added simplicity of the new Term
and Literal classes should help the solver.

Terms are now identified by 32-bit integers and come in two types,
standard names and functions. The first bit of the Term ID indicates
whether the Term is a function or not; the remaining 31 bits are used
for the term ID. Variables, sorts, arguments, and tabling will be moved
to formula level.

Literals are trivial or primitive. They are represented as a 64 bit
integer, which contains the 32 bit integers of the left- and right-hand
sides in an interleaved way (Z-order), which will allow us to use
Literals as indices of arrays. Since the right-hand side is known to
be a name, we can use its first bit to store whether the literal is an
equality.