Revised Solver::Propagate().
No lazy deletion of watchers needed anymore because when checking a
clause c that is watching f, only its watcher(s) with f are updated.
If the other watched literal has function g, then it'll be updated
later when we come to the watchers of g, and the watcher will be
removed only then.

That is, a watched literal is only updated when the loop is currently
inspecting the reference to a clause for that literal's function.