Added NogoodPredicate to Sat::Solve().
The idea is to use this predicate to dynamically add nogoods, i.e., rule
out models on the fly. In this context, a nogood is a fragment of a
model which is not admissible. A corresponding conflict clause is
generated to initiate backtrackign and avoid future models that share
the fragment.