Removed out-of-date files. Revised internal::DenseMap API.
With the ongoing rewrite, many files are completely out of date and
incompatible with the new APIs. Updated versions will be added when the
rewrite comes to an end. This particularly concerns knowledge_base.h
and action_theory.h.

I'm also far behind on unit tests.

Renamed and removed a few methods from DenseMap to avoid naming
collisions.