Added Formula::SatisfiedBy().
Only works for ground, stripped formulas.
Generates an explanation, which is a vector of literals that are
satisfied by the model and that suffice to make the formula true.
This vector is not minimal, it is just the first one found when
traversing the formula. It's supposed to serve as nogood for the
limited solver.