Fixed: Sat move semantics, using relative pointers.
Sat::ActivityCompare holds a pointer to Sat::fun_activity_.
When Sat is moved, the pointer isn't updated.
The new implementation uses pointer arithmetic to figure it out.
It's very hacky because it only works if the address of MinHeap::Less
is the same as the address of Sat::fun_queue_.

The better solution is probably to disallow moves for Sat.

The change of commit will probably be undone.