Fixed: Sat's kBase level is closed under unit propagation (in Solve()).
The bug was that after Reset(), the base level unit clauses didn't
trigger unit propagation. That's why copying Sat worked, whereas
Reset()ing it didn't.

One fix is to set trail_head_ = 0 in Reset(). To avoid re-running the
unit propagation all the time, and also to fix the invariant that every
complete level shall be closed under unit propagation, the new solution
is to run Propagate() in Solve() initially if the current level is
kBase. (This invariant could also be checked in Sat::Invariants().)